Channel tabs close when someone parts the channel, empty channels are deleted, fixed bug where messages stopped going to the proper channel
There is currently a bug that when someone joins a channel or channels, leaves a channel, and then joins another channel their client stops receiving messages. They can still send messages. Several hours of debugging and I don't know what is causing this. It won't impact the next few weeks of development (and it only occurs in a very specific scenario) and maybe I will stumble across the issue during that time so I am moving on for now.